requirements.txt
numpy
pandas
scikit-learn
pyradiomics
tensorflow==2.6.0
keras
matplotlib
seaborn
opencv-python
SimpleITK